refactor: ForwardController 삭제 및 Manual Controller 인터페이스 패키지 수정
