+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/tobe/dispatcherservlet/AppConfig.java
@@ -1,6 +1,6 @@
 package com.interface21.webmvc.servlet.mvc.tobe.dispatcherservlet;
 
-import com.interface21.webmvc.servlet.mvc.asis.Controller;
+import com.interface21.webmvc.servlet.mvc.tobe.Controller;
 import java.util.Map;
 
 public interface AppConfig {
